geolocation-API/getCurrentPosition_IDL.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<meta name="timeout" content="long" />
<title>Geolocation Test: getCurrentPosition tests</title>
<link rel="help" href="http://www.w3.org/TR/geolocation-API/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  const resetPermission = () => {
    return test_driver.set_permission({ name: "geolocation" }, "prompt");
  };
  promise_test(async (t) => {
    t.add_cleanup(resetPermission);
    await test_driver.set_permission({ name: "geolocation" }, "denied");
    const promise = new Promise((resolve, reject) => {
      navigator.geolocation.getCurrentPosition(reject, resolve);
    });
    /*
    interface GeolocationPositionError {
      const unsigned short PERMISSION_DENIED = 1;
      const unsigned short POSITION_UNAVAILABLE = 2;
      const unsigned short TIMEOUT = 3;
      readonly attribute unsigned short code;
      readonly attribute DOMString message;
    };
    */
    const error = await promise;
    assert_true(
      error instanceof GeolocationPositionError,
      "expected GeolocationPositionError"
    );
    assert_equals(GeolocationPositionError.PERMISSION_DENIED, 1);
    assert_equals(GeolocationPositionError.POSITION_UNAVAILABLE, 2);
    assert_equals(GeolocationPositionError.TIMEOUT, 3);
    assert_equals(error.code, GeolocationPositionError.PERMISSION_DENIED);
    assert_equals(
      typeof error.message,
      "string",
      "expected string for message"
    );
    await test_driver.set_permission({ name: "geolocation" }, "prompt");
  }, "User denies access, check that error callback is called.");

  promise_test(async (t) => {
    t.add_cleanup(resetPermission);
    await test_driver.set_permission({ name: "geolocation" }, "granted");
    const promise = new Promise((resolve, reject) => {
      navigator.geolocation.getCurrentPosition(resolve, reject);
    });
    const position = await promise;
    /*
    interface GeolocationPosition {
      readonly attribute GeolocationCoordinates coords;
      readonly attribute EpochTimeStamp timestamp;
    };
    */
    assert_true(position instanceof GeolocationPosition);
    assert_true(position.coords instanceof GeolocationCoordinates);
    assert_true(
      typeof position.timestamp === "number",
      "Expected timestamp to be a number"
    );
    /*
      interface GeolocationCoordinates {
        readonly attribute double accuracy;
        readonly attribute double latitude;
        readonly attribute double longitude;
        readonly attribute double? altitude;
        readonly attribute double? altitudeAccuracy;
        readonly attribute double? heading;
        readonly attribute double? speed;
      };
    */
    const coords = position.coords;
    assert_equals(
      typeof coords.accuracy,
      "number",
      "accuracy must be a number"
    );
    assert_equals(
      typeof coords.latitude,
      "number",
      "latitude must be a number"
    );
    assert_equals(
      typeof coords.longitude,
      "number",
      "longitude must be a number"
    );
    assert_equals(
      coords.altitude === null || typeof coords.altitude === "number",
      "altitude must be null or a number"
    );
    assert_equals(
      coords.altitudeAccuracy === null ||
        typeof coords.altitudeAccuracy === "number",
      "altitudeAccuracy must be null or a number"
    );
    assert_equals(
      coords.heading === null || typeof coords.heading === "number",
      "heading must be null or a number"
    );
    assert_equals(
      coords.speed === null || typeof coords.speed === "number",
      "speed must be null or a number"
    );
  }, "User allows access, check successCallback.");
</script>
